Remove unused default constructor in Decoder
The default constructor of the Decoder class was removed as it was unused and unnecessary. There was also a potential for misuse since it didn't initialize the class with the proper parameters. Removing it enforces the use of the constructor that requires necessary parameters (codec and parameters), ensuring proper initialization of the Decoder object.