Add AudioMixer to Codec
The AudioMixer has been added to the Codec, It has been created to facilitate the handling of multiple audio channels. The CMakeLists.txt file has been modified to include the corresponding new cpp and header files. The new class files (AudioMixer.cpp and AudioMixer.hpp) define the behaviour of the AudioMixer.