Add methods to count input and output in FilterGraph
This commit adds two new methods in the FilterGraph class, GetInputCount() and GetOutputCount(), to return the count of inputs and outputs respectively. These methods were added to provide an easier way to track and manage the amount of inputs and outputs in the FilterGraph.

The diff also contains reformatting of the code for improved readability and maintenance.