Change SendFrame parameter to pass-by-value
The function SendFrame in the Filter component was modified to receive a Frame parameter as pass-by-value instead of pass-by-reference. This avoids the possibility of null reference error if the caller function accidentally passes a null reference to SendFrame. This also gives freedom to manipulate the frame object within the function without changing the original object.