Change FrameBuffer type to map for better input-output correlation
The FrameBuffer type was changed from vector to map in order to match the indexes of input-output buffers and filters. This change facilitates the management of buffers when filters are added or removed. Instead of pushing to the end or emplacing back, we now emplace at a specific index, ensuring that buffer and filter indexes correspond with each other. This helps improve the overall efficiency and readability of the code.