Update BufferSource() constructor to accept AudioFrameFormat
The BufferSource() constructor was changed, now it accepts AudioFrameFormat as an argument. This was done to simplify and improve the way AudioFrameFormat is managed within the BufferSource. This change also led to the removal of multiple getter methods that were previously needed to access the sample rate, sample format, channel count and channel layout from the BufferSource. Now, this information can be directly retrieved from the provided AudioFrameFormat.