Add PTSSetter class for PTS manipulation in frames
A new class, PTSSetter, has been added to the Codec namespace. This class is responsible for handling Presentation Timestamp (PTS) within audio frames. It takes frames, checks their format and sends them into a filter graph for PTS manipulation. This change was necessary to fix synchronization issues and provide a mechanism for controlling timing in audio streams.

Changes include the creation of PTSSetter.hpp and PTSSetter.cpp with appropriate methods and implementation. These new files are also added to the CMakeLists.txt.