Add constructor to AudioFrameFormat class
A new constructor has been added to the AudioFrameFormat class in the file include/Codec/AudioFrameFormat.hpp. This constructor accepts four parameters: sampleRate, sampleFormat, an AVChannelLayout object, and an optional channelLayout parameter with a default value of 0. The AVChannelLayout object is copied into the channels member variable. This change allows for more flexible instantiation of the AudioFrameFormat objects.