Commit message: Add decoder access method to MediaStream class
Added an access method GetDecoder() to the MediaStream class, which creates and returns a Decoder object with the stream's codec and codec parameters. This simplification will help to reduce the duplication of code, as now the creation of a Decoder object can be done with a single stream method call. This change also enhances encapsulation, as it reduces the information clients need to know about the internal details of a Decoder's creation.