Add offset to SongChangedEvent struct
Added an 'offset' field to the 'SongChangedEvent' struct in 'Playlist.hpp'. This field will specify the difference in playlist index when a song is changed. This new field is then initialized in 'Playlist.cpp' at '-1' when playing the previous song and '1' when playing the next song. This is useful to keep track of the direction of song changes in the playlist.