Add a function to get playlist length
Added a new function 'Length()' to return the total number of songs in the playlist which includes previous, current and next tracks. Also, an alias 'EventReceiver' was created for readability. Moreover, the value passed to the index in 'SongAddedEvent' is now obtained via 'Length()' function, ensuring that index always represents the correct position of the song in the playlist.