Add song index to SongChangedEvent
A new field 'index' has been added to the SongChangedEvent struct in Playlist.hpp. This index shows the position of the currently playing song in the playlist. Changes have been made in Playlist.cpp to update this new field when a song change event is triggered. This allows users to know the exact position of the currently playing song.