Add PlaybackEndedEvent to Playlist module
A new event type, PlaybackEndedEvent, was introduced in our Playlist component to notify when the playback of the playlist has finished. This feature will make the playlist module more robust and allow better handling of user actions when the playlist has ended. In particular, updating user interface based on the playlist state will be made easier and more reliable.