Add methods to retrieve audio frame information
Two new methods, `GetFrameFormat()` and `GetFrameSize()`, were added to the `Playlist` class in the Codec Audio library. These methods allow external code to retrieve the format and size of the audio frames in the playlist, which was previously inaccessible. This modification provides more granular control and is useful for handling audio data more efficiently.