Add [[nodiscard]] attribute to playlist methods
The [[nodiscard]] attribute has been added to several methods in the Playlist class. This attribute ensures that the return value of these methods is not going to be ignored by warning the developer at compile time. These updates need this attribute, as neglecting their return values could lead to unintended behavior in the program.