Add frame count to MediaStream
A new method has been added in both MediaStream header and implementation files (MediaStream.hpp and MediaStream.cpp) to retrieve the total frame count from a media stream. This function is necessary for cases where we want to get the total number of frames in a media stream for processing purposes. By using this new method, it's now possible to get the number of frames without decoding all frames beforehand, thus, improving the performance in such scenarios.